SpotBugs based and JavaDoc changes

+++ SimSystem/src/main/java/mathtools/distribution/DiscreteHyperGeomDistributionImpl.java
@@ -105,8 +105,8 @@ public final class DiscreteHyperGeomDistributionImpl extends AbstractDiscreteRea
 	 * @return	Modus der Verteilung
 	 */
 	public double getMode() {
-		final int x1=(int)Math.round(Math.ceil((n+1)*(K+1)/(N+2))-1);
-		final int x2=(int)Math.round(Math.floor((n+1)*(K+1)/(N+2))-1);
+		final int x1=(int)Math.round(Math.ceil((double)(n+1)*(K+1)/(N+2))-1);
+		final int x2=(int)Math.round(Math.floor((double)(n+1)*(K+1)/(N+2))-1);
 		final double d1=getCountDensity(x1);
 		final double d2=getCountDensity(x2);
 		return (d1>d2)?x1:x2;
